requirements.txt
flask~=2.2.2
flask_jwt_extended
flask_cors
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv~=0.21.1
config
SQLAlchemy~=2.0.2